Quater4/Class-07-22-SEPTEMBER-2024-langchain-complete-app-rag-agent/pyproject.toml
[tool.poetry]
name = "class-07-22-september-2024-langchain-complete-app-rag-agent"
version = "0.1.0"
description = ""
authors = ["Muhammad Junaid <muhammadjunaid7015140@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.3.0"
langchain-core = "^0.3.5"
langchain-community = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-google-genai = "^2.0.0"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
langchain-text-splitters = "^0.3.0"
beautifulsoup4 = "^4.12.3"
faiss-cpu = "^1.8.0.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
